requirements.txt
django-suit
django-facebook
django-endless-pagination
pygments
django-image-cropping